Only invited users can register
Users are invited as soon as they are Ghost users
Ghost users are created with the register method in DatabaseParser
activateUser transforms them to "normal users" with a password
The registered = 0 field in the DB means, they are "normal users". Ghost users have this field set to 1.
Fixed some imports to the testcases run